refactor(backend): isolate Skill assignment orchestration
